# swagger-lumen/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "percobaan swagger",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.laravel-swagger-tutorial.test/",
            "description": "API server"
        }
    ],
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "List all posts",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "user"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An paged array of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "New blog post",
                "operationId": "store",
                "requestBody": {
                    "description": "Post object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PostResponse": {
                "title": "PostResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "attributes": {
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "body": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "relationships": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "title": "UserResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "attributes": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PostRequest": {
                "title": "PostRequest",
                "required": [
                    "title",
                    "body"
                ],
                "properties": {
                    "title": {
                        "description": "Class PostViewModel",
                        "type": "string"
                    },
                    "body": {
                        "description": "Class PostViewModel",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}